Move tests for pdf files to the common test file for fetch
